+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.3.10
-- **[fix]** - `ArticleImage` fix image not rendering on src change with error
+## v2.3.11
+- **[fix]** - `File` error with helper text
+- **[fix]** - `File` focusable html element